requirements.txt
apache-airflow-providers-google
# apache-airflow-providers-clickhouse
pyarrow==14.0.2
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
pymongo==4.6.1
# psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.28.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
clickhouse-driver==0.2.6